feature(template): add jest and jest-runner-eslint
