+++ generators/app/templates/README.md.template
@@ -22,10 +22,17 @@ $ npm install
 
 - Run scripts
 
-| Action                 | Usage          |
-| ---------------------- | -------------- |
-| Start development mode | `npm start`    |
-| Lint code              | `npm run lint` |
+| Action                              | Usage (npm)             | Usage (yarn)         |
+| ----------------------------------- | ----------------------- | -------------------- |
+| Start development mode              | `npm start`             | `yarn start          |
+| Run lint                            | `npm run lint`          | `yarn lint`          |
+| Start lint in watch mode            | `npm run lint:watch`    | `yarn lint:watch`    |
+| Run tests                           | `npm run jest`          | `yarn jest`          |
+| Start tests in watch mode           | `npm run jest:watch`    | `yarn jest:watch`    |
+| Run tests with code coverage        | `npm run jest:coverage` | `yarn jest:coverage` |
+| Run lint + tests                    | `npm test`              | `yarn test`          |
+| Start lint + tests in watch mode    | `npm test:watch`        | `yarn test:watch`    |
+| Run lint + tests with code coverage | `npm test:coverage`     | `yarn test:coverage` |
 
 ## Author
