+++ generators/app/templates/jest-eslint.config.js.template
@@ -1,0 +1,5 @@
+module.exports = {
+  displayName: 'lint',
+  runner: 'jest-runner-eslint',
+  testMatch: ['<rootDir>/*.js', '<rootDir>/src/**/*.js']
+};
