+++ generators/app/templates/jest-test.config.js.template
@@ -1,0 +1,7 @@
+module.exports = {
+  displayName: 'test',
+  testEnvironment: 'node',
+  testMatch: ['<rootDir>/__tests__/**/*.js'],
+  collectCoverageFrom: ['**/src/**/**.js'],
+  verbose: true
+};
