+++ generators/app/templates/jest.config.js.template
@@ -1,0 +1,6 @@
+const lint = require('./jest-eslint.config');
+const test = require('./jest-test.config');
+
+module.exports = {
+  projects: [lint, test]
+};
