--- generators/app/templates/prettierrc.template
@@ -1,4 +1,0 @@
-{
-  "arrowParens": "always",
-  "singleQuote": true
-}
